Move eth_call tests to seperate file
